Export Corner and Relationship enums as namespace properties
The enums are kept in the namespace file.
The code depending on them must be change to require the namespace
instead of the enum symbol.

This is backward compatible with code not using closure compiler.
This breaks code compiled with closure compiler (the requires must be
changed).

The solution of having Relationship in its own file in
ol/extent/relationship.js was tested but does not work: the namespace
and enum are requiring each other and while this compiles, it does not
work due to the way exportPath_ works in base.js.