Jsdoc exports: rewrite longname using module string
With modules, code is assigned in the 'exports' variable. This situation
is detected by a jsdoc plugin and the doclets are fixed to match what it
used to be when using goog.provide.

Unfortunately, this is not a simple string replacement, since:
- jsdoc has no knowledge of goog.module;
- jsdoc does some magic stuff when encountering exports variable;
- internal jsdoc '@module' support seems unusable (strange handling of dots);
- jsdoc handling of inherited docs is strange
  - not called in the plugin
  - no reference to processed file in publish.

The issue with inheritance is not addressed in this commit.