Worker based tile rendering
This POC experiments with API changes for tile preparation/rendering
inside a worker.

The main thread sends to the worker the coordinates of the tile to prepare/render.
The worker gets the source tiles, style, and prepares executor instructions / renders to an image.
The instructions / image are passed back to the main thread and populates a tile which is marked as loaded.
The layer compositor draws the tiles as usual.

Issues:
- some tiles are not hifi.